Document changed vlc format req for listening
I tried out this old tool, couldn't get it working, went spelunking in
vlc code and found the parse logic.  They've changed in a
backwards-incompatible way and now expect a URL, with some fallback for
something which looks winbindy.

So fix the README to provide a tested working example.  Again.